+++ fj-doc-base-kotlin/src/main/java/org/fugerit/java/doc/base/kotlin/dsl/Background.kt
@@ -10,8 +10,8 @@ class Background : HelperDSL.TagWithText( "background" ) {
      * Creates a new default Image instance.
      * @return the new instance.
      */
-   fun image( init: Image.() -> Unit = {} ): Image {
-       return initTag(Image(), init);
+   fun image( text: String = "", init: Image.() -> Unit = {} ): Image {
+       return initTag(Image(text), init);
    }