+++ fj-doc-base-kotlin/src/main/java/org/fugerit/java/doc/base/kotlin/dsl/Header.kt
@@ -17,8 +17,8 @@ class Header : HelperDSL.TagWithText( "header" ) {
      * Creates a new default Image instance.
      * @return the new instance.
      */
-   fun image( init: Image.() -> Unit = {} ): Image {
-       return initTag(Image(), init);
+   fun image( text: String = "", init: Image.() -> Unit = {} ): Image {
+       return initTag(Image(text), init);
    }
     /**
      * Creates a new default Phrase instance.