+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- new built-in function stringToBase64
+
 ### Changed
 
 - fj-doc-maven-plugin, init, flavour : quarkus-version set to 3.29.2 across all the modules